Remove TableObserver and table state cache
* Remove observer registry and table state cache from TableManager
* Move TableStateWatcher to its own file
* Remove TableObserver interface (only used by Manager) and replace with
  manager explicitly registering TableStateWatcher and its handler code
  with ZooCache
