+++ server/base/src/main/java/org/apache/accumulo/server/tables/TableManager.java
@@ -20,10 +20,7 @@ package org.apache.accumulo.server.tables;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
-import java.util.Collections;
 import java.util.EnumSet;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 
@@ -40,15 +37,12 @@ import org.apache.accumulo.core.fate.zookeeper.ZooUtil.NodeMissingPolicy;
 import org.apache.accumulo.core.manager.state.tables.TableState;
 import org.apache.accumulo.core.util.Pair;
 import org.apache.accumulo.core.util.tables.TableNameUtil;
-import org.apache.accumulo.core.zookeeper.ZooCache.ZooCacheWatcher;
 import org.apache.accumulo.server.ServerContext;
 import org.apache.accumulo.server.conf.store.NamespacePropKey;
 import org.apache.accumulo.server.conf.store.PropStore;
 import org.apache.accumulo.server.conf.store.TablePropKey;
 import org.apache.accumulo.server.util.PropUtil;
 import org.apache.zookeeper.KeeperException;
-import org.apache.zookeeper.WatchedEvent;
-import org.apache.zookeeper.Watcher.Event.EventType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -59,20 +53,12 @@ public class TableManager {
   private static final Logger log = LoggerFactory.getLogger(TableManager.class);
   private static final byte[] ZERO_BYTE = {'0'};
 
-  private final Set<TableObserver> observers = Collections.synchronizedSet(new HashSet<>());
-  private final Map<TableId,TableState> tableStateCache =
-      Collections.synchronizedMap(new HashMap<>());
-
   private final ServerContext context;
   private final ZooReaderWriter zoo;
 
   public TableManager(ServerContext context) {
     this.context = context;
     this.zoo = context.getZooSession().asReaderWriter();
-
-    // add our Watcher to the shared ZooCache
-    context.getZooCache().addZooCacheWatcher(new TableStateWatcher());
-    updateTableStateCache();
   }
 
   public void prepareNewNamespaceState(NamespaceId namespaceId, String namespace,
@@ -108,10 +94,6 @@ public class TableManager {
     if (!propStore.exists(propKey)) {
       propStore.create(propKey, Map.of());
     }
-  }
-
-  public TableState getTableState(TableId tableId) {
-    return tableStateCache.get(tableId);
   }
 
   public synchronized void transitionTableState(final TableId tableId, final TableState newState,
@@ -161,40 +143,25 @@ public class TableManager {
     }
   }
 
-  private void updateTableStateCache() {
-    synchronized (tableStateCache) {
-      for (String tableId : context.getZooCache().getChildren(Constants.ZTABLES)) {
-        if (context.getZooCache().get(Constants.ZTABLES + "/" + tableId + Constants.ZTABLE_STATE)
-            != null) {
-          updateTableStateCache(TableId.of(tableId));
-        }
+  public TableState getTableState(TableId tableId) {
+    TableState tState = TableState.UNKNOWN;
+    byte[] data =
+        context.getZooCache().get(Constants.ZTABLES + "/" + tableId + Constants.ZTABLE_STATE);
+    if (data != null) {
+      String sState = new String(data, UTF_8);
+      try {
+        tState = TableState.valueOf(sState);
+      } catch (IllegalArgumentException e) {
+        log.error("Unrecognized state for table with tableId={}: {}", tableId, sState);
       }
     }
-  }
-
-  public TableState updateTableStateCache(TableId tableId) {
-    synchronized (tableStateCache) {
-      TableState tState = TableState.UNKNOWN;
-      byte[] data =
-          context.getZooCache().get(Constants.ZTABLES + "/" + tableId + Constants.ZTABLE_STATE);
-      if (data != null) {
-        String sState = new String(data, UTF_8);
-        try {
-          tState = TableState.valueOf(sState);
-        } catch (IllegalArgumentException e) {
-          log.error("Unrecognized state for table with tableId={}: {}", tableId, sState);
-        }
-        tableStateCache.put(tableId, tState);
-      }
-      return tState;
-    }
+    return tState;
   }
 
   public void addTable(TableId tableId, NamespaceId namespaceId, String tableName)
       throws KeeperException, InterruptedException, NamespaceNotFoundException {
     prepareNewTableState(tableId, namespaceId, tableName, TableState.NEW,
         NodeExistsPolicy.OVERWRITE);
-    updateTableStateCache(tableId);
   }
 
   public void cloneTable(TableId srcTableId, TableId tableId, String tableName,
@@ -209,100 +176,12 @@ public class TableManager {
 
     PropUtil.setProperties(context, TablePropKey.of(tableId), propertiesToSet);
     PropUtil.removeProperties(context, TablePropKey.of(tableId), propertiesToExclude);
-
-    updateTableStateCache(tableId);
   }
 
   public void removeTable(TableId tableId) throws KeeperException, InterruptedException {
-    synchronized (tableStateCache) {
-      tableStateCache.remove(tableId);
-      zoo.recursiveDelete(Constants.ZTABLES + "/" + tableId + Constants.ZTABLE_STATE,
-          NodeMissingPolicy.SKIP);
-      zoo.recursiveDelete(Constants.ZTABLES + "/" + tableId, NodeMissingPolicy.SKIP);
-    }
-  }
-
-  public boolean addObserver(TableObserver to) {
-    synchronized (observers) {
-      synchronized (tableStateCache) {
-        to.initialize();
-        return observers.add(to);
-      }
-    }
-  }
-
-  private class TableStateWatcher implements ZooCacheWatcher {
-    @Override
-    public void accept(WatchedEvent event) {
-      if (log.isTraceEnabled()) {
-        log.trace("{}", event);
-      }
-      final String zPath = event.getPath();
-      final EventType zType = event.getType();
-
-      TableId tableId = null;
-
-      if (zPath != null && zPath.startsWith(Constants.ZTABLES + "/")) {
-        String suffix = zPath.substring(Constants.ZTABLES.length() + 1);
-        if (suffix.contains("/")) {
-          String[] sa = suffix.split("/", 2);
-          if (Constants.ZTABLE_STATE.equals("/" + sa[1])) {
-            tableId = TableId.of(sa[0]);
-          }
-        }
-        if (tableId == null) {
-          log.trace("Unhandled path {}", event);
-          return;
-        }
-      }
-
-      switch (zType) {
-        case NodeChildrenChanged:
-          // According to documentation we should not receive this event now
-          // that ZooCache is using Persistent Watchers. Not logging an error here.
-          // According to https://issues.apache.org/jira/browse/ZOOKEEPER-4475 we
-          // may receive this event (Fixed in 3.9.0)
-          break;
-        case NodeCreated:
-        case NodeDataChanged:
-          // state transition
-          if (tableId != null) {
-            TableState tState = updateTableStateCache(tableId);
-            log.debug("State transition to {} @ {}", tState, event);
-            synchronized (observers) {
-              for (TableObserver to : observers) {
-                to.stateChanged(tableId, tState);
-              }
-            }
-          }
-          break;
-        case NodeDeleted:
-          if (zPath != null && tableId != null
-              && (zPath.equals(Constants.ZTABLES + "/" + tableId + Constants.ZTABLE_STATE)
-                  || zPath.equals(Constants.ZTABLES + "/" + tableId + Constants.ZCONFIG)
-                  || zPath.equals(Constants.ZTABLES + "/" + tableId + Constants.ZTABLE_NAME))) {
-            tableStateCache.remove(tableId);
-          }
-          break;
-        case None:
-          switch (event.getState()) {
-            case Expired:
-              log.trace("Session expired; {}", event);
-              synchronized (observers) {
-                for (TableObserver to : observers) {
-                  to.sessionExpired();
-                }
-              }
-              break;
-            case SyncConnected:
-            default:
-              log.trace("Ignored {}", event);
-          }
-          break;
-        default:
-          log.warn("Unhandled {}", event);
-      }
-    }
+    zoo.recursiveDelete(Constants.ZTABLES + "/" + tableId + Constants.ZTABLE_STATE,
+        NodeMissingPolicy.SKIP);
+    zoo.recursiveDelete(Constants.ZTABLES + "/" + tableId, NodeMissingPolicy.SKIP);
   }
 
   public void removeNamespace(NamespaceId namespaceId)
