+++ server/manager/src/main/java/org/apache/accumulo/manager/Manager.java
@@ -153,7 +153,6 @@ import org.apache.accumulo.server.security.SecurityOperation;
 import org.apache.accumulo.server.security.delegation.AuthenticationTokenKeyManager;
 import org.apache.accumulo.server.security.delegation.ZooAuthenticationKeyDistributor;
 import org.apache.accumulo.server.tables.TableManager;
-import org.apache.accumulo.server.tables.TableObserver;
 import org.apache.accumulo.server.util.ScanServerMetadataEntries;
 import org.apache.accumulo.server.util.ServerBulkImportStatus;
 import org.apache.accumulo.server.util.TableInfoUtil;
@@ -184,7 +183,7 @@ import io.opentelemetry.context.Scope;
  * The manager will also coordinate log recoveries and reports general status.
  */
 public class Manager extends AbstractServer
-    implements LiveTServerSet.Listener, TableObserver, HighlyAvailableService {
+    implements LiveTServerSet.Listener, HighlyAvailableService {
 
   static final Logger log = LoggerFactory.getLogger(Manager.class);
 
@@ -1183,7 +1182,14 @@ public class Manager extends AbstractServer
 
     recoveryManager = new RecoveryManager(this, timeToCacheRecoveryWalExistence);
 
-    context.getTableManager().addObserver(this);
+    context.getZooCache().addZooCacheWatcher(new TableStateWatcher((tableId, event) -> {
+      TableState state = getTableManager().getTableState(tableId);
+      log.debug("Table state transition to {} @ {}", state, event);
+      nextEvent.event(tableId, "Table state in zookeeper changed for %s to %s", tableId, state);
+      if (state == TableState.OFFLINE) {
+        clearMigrations(tableId);
+      }
+    }));
 
     tableInformationStatusPool = ThreadPools.getServerThreadPools()
         .createExecutorService(getConfiguration(), Property.MANAGER_STATUS_THREAD_POOL_SIZE, false);
@@ -1715,20 +1721,6 @@ public class Manager extends AbstractServer
       }
     }
   }
-
-  @Override
-  public void stateChanged(TableId tableId, TableState state) {
-    nextEvent.event(tableId, "Table state in zookeeper changed for %s to %s", tableId, state);
-    if (state == TableState.OFFLINE) {
-      clearMigrations(tableId);
-    }
-  }
-
-  @Override
-  public void initialize() {}
-
-  @Override
-  public void sessionExpired() {}
 
   public Set<TableId> onlineTables() {
     Set<TableId> result = new HashSet<>();
